+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 
 - [#827](https://github.com/hyperf/hyperf/pull/827) Added a simple db component.
+- [#905](https://github.com/hyperf/hyperf/pull/905) Added twig template engine for view.
 
 ## Fixed